database/test/requirements.txt
pandas
numpy
psycopg2-binary==2.9.9
 